S3,テーブル削除の構文も取得したうえでファイル名変更
